Another significant refactor
This PR essentially rewrites the library.

- Package JS sources in src directory.
- Move `clang-format` bin entry point to bin directory.
- Simplify `bin` declaration in `package.json`.
- Make bin a thin wrapper around sources in `src`.
- Add `"use strict";` pragma to all JS sources.
- Move FS walk and source filtering out of `index` into an `fs` module.
- Add a try-catch block around `cli.run`.
- Split binpath resolution into a separate method outside of
  `format.format`.
- Remove double comparison of extenstion equality check by checking for
  `.c` and `.h` per `path.extname` API docs.
- All `async function` declarations converted to
  `const foo = async () => {}` statements.

This PR introduces a significant API change to the formatter in
`index.js`. The module exports two functions, `check` and `format` that
are constructors for formatters bound to a particular check mode and
source root. This accompanies an internal API change to split the check
and format code paths.
