+++ bin/clang-format.js
@@ -1,0 +1,9 @@
+#!/usr/bin/env/node
+
+"use strict";
+
+const main = require("../src/cli");
+
+(async () => {
+  await main();
+})();
