+++ package.json
@@ -17,10 +17,8 @@
     "Ryan Lopopolo <rjl@hyperbo.la> (https://hyperbo.la/)"
   ],
   "repository": "github:artichoke/clang-format",
-  "main": "./index",
-  "bin": {
-    "clang-format": "./cli.js"
-  },
+  "main": "./index.js",
+  "bin": "./bin/clang-format.js",
   "dependencies": {
     "commander": "^6.1.0"
   },
